+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Changelog
 * Add new error ``sdk.utils.IncorrectMetadataError``.
 * Update :command:`run-ngs-pipeline` command to check input files more vigorously.
 * Add new method :meth:`sdk.utils.compare_metadata`.
+* Add new method :meth:`api.core.get_strand`.
 
 0.8.0 (2021-11-20)
 ------------------